This commit was manufactured by cvs2svn to create tag 'nscldaq-7_4-007jumbobufs'.

+++ clients/Readout/skeleton.h
@@ -308,8 +308,8 @@ void iniscl();			/* Initialize Scaler readout.   */
 void clearevt();			/* Clear the event              */
 void clrtrig1();		/* Clear aux trigger readout    */
 void clrscl();			/* Clear scaler readout.        */
-UINT16  readevt(DAQWordBufferPtr& Buffer); /* Read an event  */
-UINT16  readscl(UINT32* pBuffer, int nScalers); /* Read scaler set   */
+UINT32  readevt(DAQWordBufferPtr& Buffer); /* Read an event  */
+UINT32  readscl(UINT32* pBuffer, int nScalers); /* Read scaler set   */
 void trig1dis();		/* Disable Aux. trigger         */
 void trig1ena();		/* Enable  Aux. trigger         */
 int  rdtrig1(UINT16* pBuffer);	/* Read out Aux. trigger.       */
